Update index.md
Sorry, I have made these changes in a separate commit. Having not realised that they would come up as separate suggestions.

Looks good, and mostly small edits from me. The main change is to correct important factual errors...

"changes in listing required packages in a Dockerfile can results in Docker rebuilding the whole image" - this is not correct. Docker only rebuilds down from from the layer where that package list is defined e.g. `COPY requirements.txt .`. It only rebuild the entire image if base image specified in `FROM ...` changes.

"Dockter deals with that problem pretty well by rebuilding all the images from the layer where changes occurred" - actually this is what Docker does, not Dockter. If you add or remove one package from `requirements.txt`, Docker will rebuild that layer and all the layers below it. Think of if as if you were writing a page of text by hand. With Docker, if you change one character in one line, you have to rewrite the whole line, and all the lines below it (in Docker terms all the layers below your change).

What Dock*t*er does is that instead of throwing out all those layers - it simply changes to them. So instead of having to rewrite the line and all the lines below it, you are just using some twink to change, add or remove characters.

We need to document this feature better with this sort of anaology and diagrams.